Added pooling to each of the dense connections in DeepDensePolicy to elimate all spatial dimensions of the convolutional filteres dramatically shrinking input to fully connected layer from ~60K to 384.
Because the actions do not include a spatial component, this should be fine.

However, this probably means that the fully connected layer has a limited ability
to learn to deal with spatial phenomena because it doesn't get explicit spatial info
so the earlier layers have to be deep enough that their receptive fields cover
the largest features (configurations) of interest. We are probably one pooling stage
short in this respect.